+++ a_stock_lib/providers/__init__.py
@@ -2,8 +2,6 @@
 
 from .validated_realtime_quotes import (
     QuoteObservation,
-    ValidatedRealtimeQuote,
-    ValidatedRealtimeQuoteProvider,
     market_session,
     validate_quote_observation,
 )
@@ -12,8 +10,6 @@ from .tushare_valuation import TushareValuationProvider
 
 __all__ = [
     "QuoteObservation",
-    "ValidatedRealtimeQuote",
-    "ValidatedRealtimeQuoteProvider",
     "TushareDividendProvider",
     "TushareFinancialProvider",
     "TushareValuationProvider",